requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
CacheControl==0.14.2
cachetools==5.5.2
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.2
Django==5.1.8
django-allauth==65.7.0
django-celery-beat==2.7.0
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
fcm-django==2.2.1
firebase-admin==6.7.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
kombu==5.5.2
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
